+++ test/fixture/define/exports-same/input.js
@@ -1,0 +1,4 @@
+define('foo', ['exports', 'bar', 'baz'], function (exports, bar, baz) {
+    exports.bar = bar
+    exports.baz = baz
+})