+++ test/fixture/define/exports-same/output.js
@@ -1,0 +1,4 @@
+import _bar from 'bar'
+import _baz from 'baz'
+export var bar = _bar
+export var baz = _baz