+++ test/fixture/define/imports-anonymous/input.js
@@ -1,0 +1,3 @@
+define('foo', ['exports', 'hello', 'world'], function (exports, hello) {
+  exports.bar = hello('world')
+})