+++ test/fixture/define/imports-anonymous/output.js
@@ -1,0 +1,3 @@
+import hello from 'hello';
+import 'world';
+export var bar = hello('world');