+++ test/fixture/define/destructuring-multiple-keys/output.js
@@ -1,0 +1,2 @@
+import { bar, baz } from 'foo';
+export default bar(baz);