+++ test/fixture/define/destructuring-named-keys/input.js
@@ -1,0 +1,3 @@
+define(['foo'], function ({ bar: baz }) {
+    return baz;
+});