+++ test/fixture/define/destructuring-named-keys/output.js
@@ -1,0 +1,2 @@
+import { bar as baz } from 'foo';
+export default baz;